refactor: image_preview controller を memory 用 / avatar 用に分割する
variant value で挙動を切り替える共通 controller を、責務ごとに分けた方が
ロジックが追いやすいため、以下の 2 つの Stimulus controller に分割する。

- memory_image_preview_controller (data-controller="memory-image-preview")
  - target: container / fileName
  - disconnect 時に [data-js-preview] な img のみ削除し、ERB 描画の既存画像は残す
- avatar_preview_controller (data-controller="avatar-preview")
  - target: container
  - connect 時に container の元 HTML を退避し、disconnect で復元する

合わせて View 3 ファイル (memories/new, memories/edit, devise/registrations/edit)
の data-controller / data-target / data-action を新しい controller 名に
書き換える。
